# public/swagger.yaml
openapi: 3.0.0
info:
  title: 'Mente Livre API'
  description: 'API REST para el sistema de gestión de terapeutas y citas de Mente Livre'
  contact:
    name: 'Mente Livre'
    email: mentelivre.pe@gmail.com
  version: 1.0.0
servers:
  -
    url: 'https://backend.mentelivre.org/'
    description: 'Servidor de Desarrollo Local'
paths:
...
  /therapists:
    get:
      tags:
        - Therapists
      summary: 'Obtener todos los terapeutas activos'
      operationId: getTherapists
      parameters:
        -
          name: all
          in: query
          description: 'Si es true, devuelve todos los terapeutas (activos e inactivos). Requiere autenticación.'
          required: false
          schema:
            type: string
            enum:
              - 'true'
              - 'false'
      responses:
        '200':
          description: 'Lista de terapeutas'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Therapist'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Therapists
      summary: 'Crear un nuevo terapeuta'
      operationId: createTherapist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TherapistCreate'
      responses:
        '201':
          description: 'Terapeuta creado exitosamente'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string }
                  data: { properties: { id: { type: string, format: uuid } }, type: object }
                type: object
        '401':
          description: 'No autenticado'
        '403':
          description: 'No autorizado'
      security:
        -
          bearerAuth: []
  '/therapists/{id}':
    get:
      tags:
        - Therapists
      summary: 'Obtener un terapeuta por ID'
      operationId: getTherapist
      parameters:
        -
          name: id
          in: path
          description: 'ID del terapeuta (UUID)'
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 'Terapeuta encontrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Therapist'
        '404':
          description: 'Terapeuta no encontrado'
    put:
      tags:
        - Therapists
      summary: 'Actualizar un terapeuta'
      operationId: updateTherapist
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TherapistUpdate'
      responses:
        '200':
          description: 'Terapeuta actualizado'
        '401':
          description: 'No autenticado'
        '403':
          description: 'No autorizado'
      security:
        -
          bearerAuth: []
  '/therapists/{therapistId}/photos':
    get:
      tags:
        - 'Therapist Photos'
      summary: 'Obtener fotos de un terapeuta'
      operationId: getTherapistPhotos
      parameters:
        -
          name: therapistId
          in: path
          required: true
          schema:
            type: string
        -
          name: photo_type
          in: query
          required: false
          schema:
            type: string
            enum:
              - profile
              - friendly
      responses:
        '200':
          description: 'Lista de fotos del terapeuta'
    post:
      tags:
        - 'Therapist Photos'
      summary: 'Crear foto para un terapeuta'
      operationId: createTherapistPhoto
      parameters:
        -
          name: therapistId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: 'Foto creada exitosamente'
        '400':
          description: 'Datos inválidos'
      security:
        -
          bearerAuth: []
  '/therapist-photos/{id}':
    get:
      tags:
        - 'Therapist Photos'
      summary: 'Obtener foto por ID'
      operationId: getTherapistPhoto
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Foto encontrada'
        '404':
          description: 'Foto no encontrada'
    put:
      tags:
        - 'Therapist Photos'
      summary: 'Actualizar foto'
      operationId: updateTherapistPhoto
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Foto actualizada exitosamente'
        '404':
          description: 'Foto no encontrada'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - 'Therapist Photos'
      summary: 'Eliminar foto'
      operationId: deleteTherapistPhoto
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Foto eliminada exitosamente'
        '404':
          description: 'Foto no encontrada'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    Therapist:
      required:
        - id
        - name
        - university
        - academic_cycle
        - hourly_rate
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Dr. Juan Pérez'
        university:
          type: string
          example: 'Universidad de Lima'
        academic_cycle:
          type: string
          example: '10mo ciclo'
        age:
          type: integer
          nullable: true
        years_experience:
          type: integer
          nullable: true
        professional_level:
          type: string
          nullable: true
        role_title:
          type: string
          example: Psicólogo/a
        specialty:
          type: string
          nullable: true
        therapeutic_approach:
          type: string
          nullable: true
        short_description:
          type: string
          nullable: true
        modality:
          type: string
          example: Online
        hourly_rate:
          type: number
          format: float
          example: 40
        is_active:
          type: boolean
        experience_topics:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            properties:
              photo_type:
                type: string
                enum:
                  - profile
                  - friendly
              photo_url:
                type: string
              photo_position:
                type: string
            type: object
        pricing:
          type: object
          additionalProperties: true
      type: object
    TherapistCreate:
      required:
        - name
        - university
        - academic_cycle
        - hourly_rate
      properties:
        name:
          type: string
        university:
          type: string
        academic_cycle:
          type: string
        age:
          type: integer
          nullable: true
        years_experience:
          type: integer
          nullable: true
        professional_level:
          type: string
          nullable: true
        specialty:
          type: string
          nullable: true
        hourly_rate:
          type: number
          format: float
        experience_topics:
          type: array
          items:
            type: string
      type: object
    TherapistUpdate:
      properties:
        name:
          type: string
        university:
          type: string
        academic_cycle:
          type: string
        age:
          type: integer
        hourly_rate:
          type: number
          format: float
        is_active:
          type: boolean
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Token JWT de autenticación. Incluir en el header: Authorization: Bearer {token}'
      bearerFormat: JWT
      scheme: bearer
...